+++ email_textutils.py
